% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_cat.r
\name{Micro_cat_label}
\alias{Micro_cat_label}
\title{helper add label in the micro - preprocessing stage in R do file}
\usage{
Micro_cat_label(var, var.name, add.total = FALSE)
}
\arguments{
\item{var}{variable from the microdataset to manipulate.}

\item{var.name}{character, variable name to manipulate in input.}

\item{add.total}{internal attribute to add total during the processing.}

\item{to}{character, variable name to return.}
}
\description{
not longer used
}
\examples{
## Not run:

## End(**Not run**)
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{preprocessing}
