% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_check_variable.r
\name{Micro_check_variable}
\alias{Micro_check_variable}
\title{helper to check dta columns and value present over times series}
\usage{
Micro_check_variable(my_pattern = NULL, STORE = FALSE)
}
\arguments{
\item{my_pattern}{character for selection of dataset with pattern choose.}

\item{STORE}{TRUE or FALSE, if TRUE will combine all variables by dataset in a single file.}
}
\description{
Support for recoding variable, check \code{?car::Recode}, for labelling variable
}
\examples{
## Not run:

 # first consolidate all existing variables
Micro_check_variable(STORE = TRUE)

# then look at variable pattern needed:
Micro_check_variable(my_pattern = "job1_isci3")




## End(**Not run**)
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
