% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_convert.r
\name{Micro_convert}
\alias{Micro_convert}
\title{helper to convert fil in dta or sav}
\usage{
Micro_convert(in_file = NULL, country = NULL, source = NULL,
  time = NULL, ori = TRUE, encoding = NULL)
}
\arguments{
\item{in_file}{dta or sav file path for loading.}

\item{country}{character for selection of country (iso3 code), default NULL.}

\item{source}{character for selection of source (as ilo micro spelling, ie. LFS), default NULL.}

\item{time}{, character, time, use to rebuilt a specific dataset, default NULL means all time will be takeinto account.}

\item{ori}{, boolean fefault  = TRUE, look inside ORI folder to detect sav file to convert in dta}
}
\description{
faster method to reencode file in correct format
}
\details{
Support
}
\examples{
## Not run:

Micro_convert(in_file = 'J:/COMMON/STATISTICS/DPAU/MICRO/ARG/QLFS/2003Q3/ORI/Individual_t303.sav')

Micro_convert(country = 'ARG', source = 'LFS', time = '2010', encoding  = 'latin1')

## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
