% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_create_ddi.r
\name{Micro_create_ddi}
\alias{Micro_create_ddi}
\title{helper to prepare summary file on root repo}
\usage{
Micro_create_ddi(x, ref_ddi)
}
\description{
helper to prepare summary file on root repo
}
\examples{
## Not run:
Micro_file_ready()

## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
