% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_file_ready.r
\name{Micro_file_ready}
\alias{Micro_file_ready}
\title{helper to prepare summary file on root repo}
\usage{
Micro_file_ready()
}
\description{
helper to prepare summary file on root repo
}
\examples{
## Not run:
Micro_file_ready()

## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
