% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_get_workflow.r
\name{Micro_get_workflow}
\alias{Micro_get_workflow}
\title{helper to work with micro data preprocessing workflow}
\usage{
Micro_get_workflow(
  ref_area = NULL,
  source = NULL,
  time = NULL,
  timefrom = NULL,
  timeto = NULL
)
}
\arguments{
\item{ref_area}{character vector for selection of country (iso3 code), default NULL, all.}

\item{source}{character vector for selection of source (as ilo micro spelling, ie. LFS), default NULL, all.}

\item{time}{character vector for selection of time, default NULL, all.}
}
\description{
helper to work with micro data preprocessing workflow
}
\examples{
## Not run:
## get all :

Micro_get_workflow()

## get only one ref_area one source:



Micro_get_workflow(ref_area = 'CRI', source = 'LFS')

## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
