% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_backup_cmd.r, R/Micro_cat.r,
%   R/Micro_get_workflow.r, R/Micro_process.r
\name{Micro_backup_cmd}
\alias{Micro_backup_cmd}
\alias{Micro_cat_label}
\alias{Micro_cat_recode}
\alias{Micro_get_workflow}
\alias{Micro_process}
\title{backup do for Master files}
\usage{
Micro_backup_cmd()

Micro_cat_label(var, var.name, add.total = FALSE)

Micro_cat_recode(var, var.name, to)

Micro_get_workflow(country = NULL, source = NULL, time = NULL)

Micro_process(country = NULL, source = NULL, time = NULL,
  indicator = NULL, ktest = c(max = 15, min = 5, threshold = 0.2),
  saveCSV = TRUE, ilo_time = NULL, delete = FALSE, collection = "YI",
  keep_count = TRUE, query = NULL, print_ind = TRUE, validate = FALSE,
  consolidate = "1")
}
\arguments{
\item{var}{variable from the microdataset to manipulate.}

\item{var.name}{character, variable name to manipulate in input.}

\item{add.total}{internal attribute to add total during the processing.}

\item{to}{character, variable name to return.}

\item{country}{character vector for selection of country (iso3 code), default NULL, all.}

\item{source}{character vector for selection of source (as ilo micro spelling, ie. LFS), default NULL, all.}

\item{time}{character vector for selection of time, default NULL, all.}

\item{indicator}{allow to select a specific indicatr see example, default NULL means that all indicator available will be generate from the microdataset.}

\item{ktest}{numeric vector, thresholds define to determine whether an indicator is unrealable or not, see section, default c(max = 30, min = 5, threshold = 0.2).}

\item{saveCSV}{save result in ilostat csv format on the directory of the path, default FALSE, if TRUE tes on scope, indicator, version is done.}

\item{ilo_time}{force microdataset ilo_time variable to a new define value, ideal to move quarterly data to yearly, default NULL.}

\item{delete}{thresholds define to determine whether an indicator is unrealable or not, see section.}

\item{collection}{map in another collection.}

\item{keep_count}{expert only.}

\item{query}{expert only.}

\item{print_ind}{print indicator processed, default = TRUE.}

\item{validate}{only ready file on workflow.}

\item{path}{dta or zip file path for loading ilo microdataset pre precessed, default = NULL.}

\item{country}{character for selection of country (iso3 code), mandatory if path is not set.}

\item{source}{character for selection of source (as ilo micro spelling, ie. LFS), mandatory if path is not set.}

\item{time}{, character, time, use for a specific dataset, default NULL, mandatory if path is not set.}

\item{test}{expert only.}
}
\description{
not longer used

not longer used

helper to work with micro data preprocessing workflow

faster method to recode, order, add label of categorical variables
}
\details{
Support for recoding variable, check \code{?car::Recode}, for labelling variable
}
\examples{
## Not run:

## End(**Not run**)
## Not run:

## End(**Not run**)
## Not run:
## get all :

Micro_get_workflow()

## get only one country one source:



Micro_get_workflow(country = 'CRI', source = 'LFS')

## End(**Not run**)
## Not run:

# test that
 Micro_process(country = 'ZAF', source = 'LFS', time = '2016Q1', saveCSV = TRUE, validate = TRUE)

# process the all time series validate in the workflow
 Micro_process(country = 'ZAF', source = 'LFS', print_ind = FALSE, saveCSV = TRUE, validate = TRUE)


### play with parameters: only for expert
# too much details, no value could be publish

Micro_process(indicator = 'UNE_TUNE_SEX_AGE_EDU_NB')

# change the releability threshold:

Micro_process(indicator = 'UNE_TUNE_SEX_AGE_EDU_NB', ktest = c(max = 30, min = 5, threshold = 0.2))

# or move to yearly data 

Micro_process(indicator = 'UNE_TUNE_SEX_AGE_EDU_NB', ilo_time = '2015')

# check why data have been discarded (display sample_count and table_test variable)

Micro_process(indicator = 'UNE_TUNE_SEX_AGE_EDU_NB', delete  = FALSE)

# use your own dataset

Micro_process(path = 'C:/user/../LKA_LFS_2013.dta')

# run the entire process
Micro_process(collection = 'STI', print_ind = FALSE, validate = TRUE, consolidate = '123')
# run the entire process for a country only
Micro_process(country ='ZAF', source = 'LFS', collection = 'STI', print_ind = FALSE, validate = TRUE, consolidate = '123')

## End(**Not run**)
}
\author{
ILO / bescond

ILO / bescond

ILO / bescond

ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{preprocessing}
\keyword{processing}
