% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_query.r
\name{Micro_query_create}
\alias{Micro_query_create}
\title{helper to prepare query}
\usage{
Micro_query_create(query = NULL)
}
\arguments{
\item{path}{dta or zip file path for loading ilo microdataset pre precessed, default = NULL.}
}
\description{
faster method to recode, order, add label of categorical variables
}
\examples{
## Not run:

# process the all time series validate in the workflowflow
 Micro_process(ref_area = 'ZAF', source = 'QLFS', saveCSV = TRUE, validate = TRUE, time = '2016Q1', print_ind = TRUE)


# run the entire process for one country only (go for publication on ilostat)
Micro_process(validate = TRUE, consolidate = '123', PUB = TRUE, skip = 'EULFS', ref_area ='ZAF', source = 'LFS')


# you could now rebuild, process 1 for a single time, then process 23
# use mainly for adding oe dataset to a series yet publish

Micro_query_create(query = "MICRO_TEST")




## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing,}
\keyword{query}
