% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publish_micro.r
\name{Publish_micro}
\alias{Publish_micro}
\title{helper to calculate ilo indicator (volume and ratio) for one time period only}
\usage{
Publish_micro(myref_area, mysource, mytime)
}
\arguments{
\item{myref_area}{character for selection of country (iso3 code), mandatory if path is not set.}

\item{mysource}{character for selection of source (as ilo micro spelling, ie. LFS), mandatory if path is not set.}

\item{mytime}{, character, time, use for a specific dataset, default NULL, mandatory if path is not set.}

\item{path}{dta or zip file path for loading ilo microdataset pre precessed, default = NULL.}
}
\description{
faster method to recode, order, add label of categorical variables
}
\examples{
## Not run:
# you could now rebuild, process 1 for a single time, then process 23
# use mainly for adding oe dataset to a series yet publish

Publish_micro("CAN", "LFS", "2020M11")




## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
