% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iloMicro.r
\name{iloMicro}
\alias{iloMicro}
\title{iloMicro: Tools kit for ilo microdata processing of household / labour force survey}
\description{
description of the project here
}
\details{
Targets:

\itemize{
\item process lfs microdatafile
\item prepare summary table for current file available
\item helper to convert sav to dta and viseversa
}
}
