% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_check_dta_series.r
\name{Micro_check_dta_series}
\alias{Micro_check_dta_series}
\alias{Micro_check_dta_series_view}
\title{helper to check dta columns and value present over times series}
\usage{
Micro_check_dta_series(country = NULL, source = NULL, time = NULL,
  validate = TRUE)

Micro_check_dta_series_view(df)
}
\arguments{
\item{country}{character for selection of country (iso3 code), mandatory if path is not set.}

\item{source}{character for selection of source (as ilo micro spelling, ie. LFS), mandatory if path is not set.}

\item{time}{, character, time, use for a specific dataset, default NULL, mandatory if path is not set.}

\item{validate}{, use only ready files, default TRUE}
}
\description{
Support for recoding variable, check \code{?car::Recode}, for labelling variable
}
\examples{
## Not run:


X <- Micro_check_dta_series(country = 'ECU', source = 'LFS')

## End(**Not run**)
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
