% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_check_dta_series_view.r
\name{Micro_check_dta_series_view}
\alias{Micro_check_dta_series_view}
\title{helper to check dta columns and value present over times series}
\usage{
Micro_check_dta_series_view(ref_area, source, freq = "A")
}
\arguments{
\item{ref_area}{character for selection of country (iso3 code), mandatory if path is not set.}

\item{source}{character for selection of source (as ilo micro spelling, ie. LFS), mandatory if path is not set.}
}
\description{
Support for recoding variable, check \code{?car::Recode}, for labelling variable
}
\examples{
## Not run:

 be sure having yet run:

 Micro_check_dta_series(ref_area = 'IND', source = 'PLFS')

 then run : 

Micro_check_dta_series_view(ref_area = 'IND', source = 'PLFS', freq = 'M')

## End(**Not run**)
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
