% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_load.r
\name{Micro_load}
\alias{Micro_load}
\title{helper to load ilo.dta file for the processing process}
\usage{
Micro_load(path = NULL, query = "ILOSTAT", asFactor = FALSE, tpl = TRUE)
}
\arguments{
\item{path}{dta or zip file path for loading ilo microdataset pre precessed, default 'tpl' means that predifine example will be load, if NULL load only ilo framework.}

\item{query}{indicator definition subset}

\item{asFactor}{loading microdataset with code and label (as.factor) if TRUE only numeric code.}

\item{tpl}{default = TRUE return micro template Variable, Mapping_indicator, Mapping_classif.}
}
\description{
dataset could be manipulate during the load file base on the workflow arguments
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
