% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_process.r
\name{Micro_process}
\alias{Micro_process}
\title{helper to calculate ilo indicator (volume and ratio) for one time period only}
\usage{
Micro_process(ref_area = NULL, source = NULL, time = NULL,
  indicator = NULL, ktest = c(max = 15, min = 5, threshold = 0.334),
  saveCSV = TRUE, ilo_time = NULL, delete = FALSE, collection = "YI",
  keep_count = TRUE, query = NULL, print_ind = TRUE, validate = FALSE,
  consolidate = "1", skip = NULL, timefrom = NULL, timeto = NULL,
  output = "ilostat")
}
\arguments{
\item{ref_area}{character for selection of country (iso3 code), mandatory if path is not set.}

\item{source}{character for selection of source (as ilo micro spelling, ie. LFS), mandatory if path is not set.}

\item{time}{, character, time, use for a specific dataset, default NULL, mandatory if path is not set.}

\item{indicator}{allow to select a specific indicatr see example, default NULL means that all indicator available will be generate from the microdataset.}

\item{ktest}{numeric vector, thresholds define to determine whether an indicator is unrealable or not, see section, default c(max = 30, min = 5, threshold = 0.334).}

\item{saveCSV}{save result in ilostat csv format on the directory of the path, default FALSE, if TRUE tes on scope, indicator, version is done.}

\item{ilo_time}{force microdataset ilo_time variable to a new define value, ideal to move quarterly data to yearly, default NULL.}

\item{delete}{thresholds define to determine whether an indicator is unrealable or not, see section.}

\item{collection}{map in another collection.}

\item{keep_count}{expert only.}

\item{query}{expert only.}

\item{print_ind}{print indicator processed, default = TRUE.}

\item{validate}{only ready file on workflow.}

\item{skip}{for excuding a source as character}

\item{timefrom}{query to reduce time as from starting year timefrom.}

\item{timeto}{query to reduce time as from ending year timeto.}

\item{output}{if one year one specific indicator, return in ilo template format. Default 'ilostat'}

\item{path}{dta or zip file path for loading ilo microdataset pre precessed, default = NULL.}

\item{test}{expert only.}
}
\description{
faster method to recode, order, add label of categorical variables
}
\examples{
## Not run:

# test that
 Micro_process(ref_area = 'ZAF', source = 'QLFS', time = '2016Q1')

# process the all time series validate in the workflowflow
 Micro_process(ref_area = 'ZAF', source = 'QLFS', print_ind = FALSE, saveCSV = TRUE, validate = TRUE)


# run the entire process for one country only
Micro_process(collection = 'STI', print_ind = FALSE, validate = TRUE, consolidate = '123',ref_area ='ZAF', source = 'LFS')
# run the entire process
Micro_process(collection = 'STI', print_ind = FALSE, validate = TRUE, saveCSV = FALSE, consolidate = '123', skip = 'EULFS')

## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
