% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_process.r
\name{Micro_process}
\alias{Micro_process}
\title{helper to calculate micro indicator (volume and ratio) for ilo data and query}
\usage{
Micro_process(
  ref_area = NULL,
  source = NULL,
  time = NULL,
  timefrom = NULL,
  timeto = NULL,
  consolidate = "1",
  validate = FALSE,
  PUB = FALSE,
  ktest = c(max = 15, min = 5, threshold = 0.334),
  saveCSV = FALSE,
  QUERY = "ILOSTAT",
  ICLS = "13",
  CMD = FALSE,
  TEST_SCOPE = FALSE
)
}
\arguments{
\item{ref_area}{character for selection of country (iso3 code), mandatory if path is not set.}

\item{source}{character for selection of source (as ilo micro spelling, ie. LFS), mandatory if path is not set.}

\item{time}{, character, time, use for a specific dataset, default NULL, mandatory if path is not set.}

\item{timefrom}{query to reduce time as from starting year timefrom.}

\item{timeto}{query to reduce time as from ending year timeto.}

\item{consolidate}{default "1", or "123".}

\item{validate}{only ready file on workflow.}

\item{PUB}{if TRUE will process only dataset ready for ilostat publication (ie. workflow on_ilostat = 'Yes') else all ready/published. Default FALSE.}

\item{ktest}{numeric vector, thresholds define to determine whether an indicator is unrealable or not, see section, default c(max = 30, min = 5, threshold = 0.334).}

\item{saveCSV}{save result in ilostat csv format on the directory of the path, default FALSE, if TRUE tes on scope, indicator, version is done.}

\item{QUERY}{process only indicator dedicate to a special query, not the entire set of indicators (ie. indicator template columns query_folder set to not null). Default "ILOSTAT".}

\item{ICLS}{"13" as default, if "19" take ILO19.dta when available, use only for QUERY.}

\item{CMD}{, default FALSE, if true return str with cmd line for all Published and Ready Country/Source}
}
\description{
faster method to recode, order, add label of categorical variables. revision: 04/01/2022
}
\examples{
## Not run:

 Micro_process(saveCSV = TRUE, validate = TRUE, ref_area = 'ZAF', source = 'LFS', time = '2016Q1')


Micro_process(validate = TRUE, consolidate = '123', PUB = TRUE, ref_area ='ZAF', source = 'LFS')


# you could now rebuild, process 1 for a single time, then process 23
# use mainly for adding oe dataset to a series yet publish

Publish_micro("CAN", "LFS", "1980M11")

 # run all in parallel

cmd <- Micro_process(CMD = TRUE)
 run_parallel_ilo(cmd)
## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
