% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_rebuild_series.r
\name{Micro_rebuild_series}
\alias{Micro_rebuild_series}
\title{helper to work with micro data preprocessing workflow}
\usage{
Micro_rebuild_series(master_id = "all", ref_area, source, time = NULL,
  clean = TRUE, timefrom = NULL, timeto = NULL)
}
\arguments{
\item{master_id}{character ie. '1' for working with master_1 and copy_1. Default = 'all', could be also '12' for master 1 and 2.}

\item{ref_area}{character for selection of country (iso3 code), mandatory.}

\item{source}{character for selection of source (as ilo micro spelling, ie. LFS), mandatory.}

\item{time}{, character, time use to rebuilt a specific dataset, default NULL, means all time will be takeinto account, option.}

\item{clean, }{if TRUE delete files contained in the rebuild folder (not ORI repo), default TRUE.}

\item{timefrom}{query to reduce time as from starting year timefrom.}

\item{timeto}{query to reduce time as from ending year timeto.}
}
\description{
From Master_x rebuild time series by copy readMe, do and fwk files, then run do file and get ILO.dta and FULL.dtaand
}
\examples{
## Not run:

## some example

	Micro_rebuild_series(master_id = '1', ref_area = 'PHL', source = 'LFS', time = '2015Q3')

 Micro_rebuild_series(master_id = '1', ref_area = 'PHL', source = 'LFS') # Master 1

 Micro_rebuild_series(master_id = '12', ref_area = 'PHL', source = 'LFS') # Master 1 and Master2

 Micro_rebuild_series(ref_area = 'PHL', source = 'LFS') # all Masters

## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
