% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_rebuild_series.r
\name{Micro_rebuild_series}
\alias{Micro_rebuild_series}
\title{helper to work with micro data preprocessing workflow}
\usage{
Micro_rebuild_series(master_id = 1, country, source, ori_pattern,
  time = NULL, clean = TRUE)
}
\arguments{
\item{master_id}{ie. 1 for working with master_1 and copy_1. Default = 1.}

\item{country}{character for selection of country (iso3 code), mandatory.}

\item{source}{character for selection of source (as ilo micro spelling, ie. LFS), mandatory.}

\item{ori_pattern}{character pattern always found in the names of the original dta files, ie '_EnemduBDD', that will be detect in order to replace global input file on the do file. mandatory}

\item{time}{, character, time use to rebuilt a specific dataset, default NULL, means all time will be takeinto account, option.}

\item{clean, }{if TRUE delete files contained in the rebuild folder (not ORI repo), default TRUE.}
}
\description{
From Master_x rebuild time series by copy readMe, do and fwk files, then run do file and get ILO.dta and FULL.dtaand
}
\examples{
## Not run:

## some example
	Micro_rebuild_series(master_id = 2, country = 'ARG', source = 'QLFS', ori_pattern = 'individual_')

 Micro_rebuild_series(master_id = 1, country = 'BRA', source = 'QLFS', ori_pattern = 'pnadc_', time = '2015Q3')

 Micro_rebuild_series(master_id = 2, country = 'ECU', source = 'LFS', ori_pattern = '_EnemduBDD', time = '2013Q4')
 Micro_rebuild_series(master_id = 1, country = 'GTM', source = 'LFS', ori_pattern = 'Personas', time = '2015Q4')

 Micro_rebuild_series(master_id = 1, country = 'IDN', source = 'LFS', ori_pattern = 'IDN_', time = '2014Q3')

 Micro_rebuild_series(master_id = 1, country = 'MEX', source = 'QLFS', ori_pattern = 'sdemt')

 Micro_rebuild_series(master_id = 1, country = 'ZAF', source = 'LFS', ori_pattern = 'lfs')

## End(**Not run**)
}
\author{
ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{processing}
